requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
mysql-connector-python==8.0.32
protobuf==3.20.3
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
